changed data.frame's to data frames
Since no possession is implied no need for the apostrophe. In the docs they use 'data frame' and  'data frames' without the period unless it's in code tags, so we have a precedent for writing it without a period. P.S. this whole R for cats site is great, thanks! I'm requiring my students to read it.